requirements.txt
fastapi
uvicorn[standard]
langchain-community
langchain
openai
psycopg2-binary
pdfplumber
tiktoken
python-dotenv
pydantic-settings
pinecone-client
gunicorn
langchain-pinecone
python-multipart
httpx
pytest
pytest-asyncio
loguru
requests
python-docx
tqdm
# Free alternatives
huggingface-hub
sentence-transformers
transformers
torch
# Optional: for Ollama
ollama